mdb-6-creating-a-link-for-a-movie-detail
dividing movie page and tv show page.
adding image placeholder for movie that doesn't have image
