+++ src/pages/Home.jsx
@@ -1,12 +1,10 @@
 import React from 'react'
 import { endURL } from '../api/request'
 import Content from '../component/Content'
-// import Header from '../component/Header'
 
 function Home() {
   return (
     <>
-      {/* <Header /> */}
       <Content title='trending' baseURL={endURL.trending} />
       <Content title='upcoming' baseURL={endURL.upcoming} />
       <Content title='popular' baseURL={endURL.popular} />
